create seacat-auth folder
