+++ src/index.js
@@ -9,7 +9,6 @@ export { DataTableFilter2, DataTableAdvFilterSingleValue2, DataTableAdvFilterMul
 export { Spinner } from './components/Spinner';
 export { default as ControlledSwitch } from './components/ControlledSwitch';
 export { default as UncontrolledSwitch } from './components/UncontrolledSwitch';
-export { Credentials } from './components/Authz/Credentials';
 export { CellContentLoader, ChartLoader } from './components/ContentLoader';
 export { DateTime } from './components/DateTime/DateTime.jsx';
 export { DateTimeRelative } from './components/DateTime/DateTimeRelative.jsx';
@@ -26,11 +25,6 @@ export { itemExtensionHandler } from './utils/itemExtensionHandler.js';
 export { removeFileExtension } from './utils/removeFileExtension.js';
 export { deepMerge } from './utils/deepMerge.jsx';
 export { problemMarkers } from './utils/monaco/problemMarkers.jsx';
-export { ButtonWithAuthz } from './components/Authz/ButtonWithAuthz';
-export { LinkWithAuthz } from './components/Authz/LinkWithAuthz';
-export { ControlledSwitchWithAuthz } from './components/Authz/ControlledSwitchWithAuthz';
-export { UncontrolledSwitchWithAuthz } from './components/Authz/UncontrolledSwitchWithAuthz';
-export { isAuthorized } from './components/Authz/utils/isAuthorized.js';
 export { ResultCard } from './components/ResultCard/ResultCard';
 export { AdvancedCard } from './components/AdvancedCard/AdvancedCard.jsx';
 export { AsabReactJson } from './components/AsabReactJson/AsabReactJson.jsx';
